changes in the DB

+++ power_app/models.py
@@ -18,7 +18,7 @@ class Usuario(models.Model):
     nombre = models.CharField(max_length=50)
     apellido = models.CharField(max_length=50)
     sexo = models.CharField(max_length=10, choices=SEX_CHOICES)
-    DNI = models.IntegerField(unique=True)
+    DNI = models.BigIntegerField(unique=True)
     celular = models.CharField(max_length=20, blank=True, null=True)
     pago = models.DateField(default=timezone.now, null=False, blank=False)
     vencimiento = models.DateField(
